Remove common feature in uint
