+++ uint/Cargo.toml
@@ -18,11 +18,10 @@ crunchy = { version = "0.2", default-features = true }
 [features]
 default = ["std"]
 std = ["byteorder/std", "rustc-hex/std", "crunchy/std"]
-common = []
 
 [[example]]
 name = "modular"
 
 [[test]]
 name = "uint_tests"
-required-features = ["std", "common"]
+required-features = ["std"]
